一旦いろいろ確認

+++ Projects/SoLib/Math/Angle.h
@@ -88,12 +88,12 @@ namespace Angle {
 /// @param val 度数法
 /// @return 弧度法
 inline float operator""_deg(long double val) {
-	return static_cast<float>(val) * Angle::Dig2Rad;
+	return static_cast<float>(val * Angle::Dig2Rad);
 }
 
 /// @brief 度数法の値を弧度法に変換
 /// @param val 度数法
 /// @return 弧度法
 inline float operator""_deg(size_t val) {
-	return static_cast<float>(val) * Angle::Dig2Rad;
+	return static_cast<float>(static_cast<double>(val) * Angle::Dig2Rad);
 }
